// src/app/models/IMovie.ts
export interface IMovie {
  backdrop_path: string;
  createdAt: string;
  genre: [string];
  id: string;
  movie_id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  updatedAt: string;
  video: string;
  vote_average: number;
  vote_count: number;
  __v: number;
  _id: string;
}